third_party/WebKit/LayoutTests/editing/selection/image-before-linebreak.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable id="div"><img src="../resources/abe.png"><br>foo</div>',
    selection => {
        // Select the line break
        var e = selection.document.getElementById("div");
        selection.setBaseAndExtent(e, 1, e, 2);
    },
    '<div contenteditable id="div"><img src="../resources/abe.png">^<br>|foo</div>'),
    'Only the line break after the image should be selected.');
</script>
